Move function not api in Controller
