+++ docs/api.md
@@ -2,11 +2,8 @@
 | Route         | Type   |  Description     |  Args     | Return     |
 |-------------|:------:|------------------|----------|----------|
 | api/tasks     | GET    | Get all project tasks  | *long* projectID | *Array\<Task>*
-| api/tasks/graph    | GET    | Get data for graph  | *long* taskID | *TaskGraphWrapper*
 | api/tasks     | POST    | Create task  | *TaskWrapper* task  | *UpdateTaskResult*
-| api/tasks/delete     | GET    | Delete  task | *long* taskID | ACK
-| api/tasks/approve     | GET    | Approve task  | *long* taskID | ACK
-| api/tasks/deny     | GET    | Deny task  | *long* taskID | ACK
+
 |
